requirements.txt
streamlit==0.71.0
pyngrok==4.1.3
tensorflow==2.3.1
keras==2.4.3
pillow==8.0
h5py==2.10.0